adding containable behaviour class for streamline
Model behaviour allows you to filter and limit model find operations.
Containable allows you to streamline and simplify operations on your
model bindings. It works by temporarily or permanently altering the
associations of your models. It does this by using supplied the
containments to generate a series of bindModel and unbindModel calls.

To attach behaviour on the fly:
$this->Post->Behaviors->load('Containable');